Reorganize controls, export react controllers
